requirements.txt
torch==2.2.1
torchvision==0.17.1
diffusers==0.27.2
pillow==10.2.0
numpy==1.24.4
tqdm==4.66.2
huggingface-hub==0.21.4
accelerate==0.28.0
transformers==4.40.2
xformers==0.0.25
datasets==2.19.2